% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\docType{data}
\name{dict}
\alias{dict}
\title{Diccionario para los datos de las encuestas ENGIH 2018 (Banco Central)}
\format{
\link{list}
}
\source{
\url{https://github.com/endomer-py/engihr/tree/main/Files/Cuestionarios/}
}
\usage{
dict
}
\description{
Contiene las etiquetas tanto para las variables como para los valores de estas.
}
\keyword{datasets}
