% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egi_Labels.R
\name{egi_set_labels}
\alias{egi_set_labels}
\alias{egi_setLabels}
\title{Asigna etiquetas de datos a las variables especificadas
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
egi_set_labels(tbl, dict = engihr::dict, vars = NULL)

egi_setLabels(tbl, dict = engihr::dict, vars = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{dict}{\link{list}: Lista de etiquetas a asignar a las variables.}

\item{vars}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}
}
\value{
Los datos introducidos en el argumento \code{tbl} pero con etiquetas de datos
}
\description{
Asigna etiquetas de datos a las variables especificadas
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Esta función permite asignar etiquetas de datos a las variables de un data.frame
o equivalente, a partir del diccionario de la encuesta.
}
\examples{
\dontrun{
  engih <- data.frame(ZONA = c(1, 2))
  str(engih)
  str(egi_set_labels(engih))
}
}
\seealso{
Etiquetas de datos \code{vignette("labeler", package = "labeler")}
}
