% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egi_Labels.R
\name{egi_use_labels}
\alias{egi_use_labels}
\alias{egi_useLabels}
\title{Utiliza las etiquetas de datos de una variable
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
egi_use_labels(tbl, dict = engihr::dict, vars = NULL)

egi_useLabels(tbl, dict = engihr::dict, vars = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{dict}{\link{list}: Lista de etiquetas a asignar a las variables.}

\item{vars}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}
}
\value{
Los datos suministrados en el argumento \code{tbl}, pero en lugar de
valores utilizando las etiquetas de datos correspondientes
}
\description{
Utiliza las etiquetas de datos de una variable
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
  engih <- data.frame(ZONA = c(1, 2))
  engih
  egi_use_labels(engih)
}
}
\seealso{
Etiquetas de datos \code{vignette("labeler", package = "labeler")}
}
